scraper/requirements.txt
curl-cffi==0.7.4
beautifulsoup4==4.12.3
lxml==5.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
